Check if new reports can be created during first migration

+++ db/migrate/20140525190034_create_default_report_configuration.rb
@@ -4,26 +4,31 @@
 
 class CreateDefaultReportConfiguration < ActiveRecord::Migration
   def up
-  	ReportConfiguration.new do |rc|
-  		rc.image = File.open(Rails.root.join("public/images/logoUFF.jpg").to_s)
-  		rc.name = "Padrão"
-  		rc.show_sapos = false
-  		rc.text = "UNIVERSIDADE FEDERAL FLUMINENSE
+
+    if ActiveRecord::Base.connection.table_exists? CarrierWave::Storage::ActiveRecord::ActiveRecordFile.table_name
+      ReportConfiguration.new do |rc|
+        rc.image = File.open(Rails.root.join("public/images/logoUFF.jpg").to_s)
+        rc.name = "Padrão"
+        rc.show_sapos = false
+        rc.text = "UNIVERSIDADE FEDERAL FLUMINENSE
 INSTITUTO DE COMPUTAÇÃO
 PROGRAMA DE PÓS-GRADUAÇÃO EM COMPUTAÇÃO"
-  		rc.use_at_grades_report = true
-  		rc.use_at_report = true
-  		rc.use_at_schedule = true
-  		rc.use_at_transcript = true
-  		rc.order = 1
-      rc.scale = 0.4
-      rc.x = 13
-      rc.y = 13
-  	end.save!
+        rc.use_at_grades_report = true
+        rc.use_at_report = true
+        rc.use_at_schedule = true
+        rc.use_at_transcript = true
+        rc.order = 1
+        rc.scale = 0.4
+        rc.x = 13
+        rc.y = 13
+      end.save!
+    end
   end
 
   def down
-  	rc = ReportConfiguration.find_by_name("Padrão")
-    rc.destroy unless rc.nil?
+    if ActiveRecord::Base.connection.table_exists? CarrierWave::Storage::ActiveRecord::ActiveRecordFile.table_name
+      rc = ReportConfiguration.find_by_name("Padrão")
+      rc.destroy unless rc.nil?
+    end
   end
 end
